SMCsamplers.FFBS — Function

Xdraws = FFBS(U, Y, A, B, C, Σₑ, Σₙ, μ₀, Σ₀, nSim = 1)

Forward filtering and backward sampling from the joint smoothing posterior p(x1,...xT | y1,...,yT) of the state space model:

yₜ = Cxₜ + εₜ, εₜ ~ N(0,Σₑ) Measurement equation

xₜ = Axₜ₋₁+ Buₜ + ηₜ, ηₜ ~ N(0,Σₙ) State equation

where

xₜ is the n-dim state

uₜ is the m-dim control

yₜ is the k-dim observed data.

The observed data observations are the rows of the T×k matrix Y The control signals are the rows of the T×m matrix U μ₀ and Σ₀ are the mean and covariance of the initial state vector x₀. A, C, Σₑ and Σₙ can be deterministically time-varying by passing 3D arrays of size n×n×T.

Note: If nSim == 1, the returned Xdraws is matrix, otherwise it is a 3D array of size T×n×nIter.

SMCsamplers.FFBSx — Function

FFBSx(U, Y, A, B, C, ∂C, Cargs, Σₑ, Σₙ, μ₀, Σ₀)

Forward filtering and backward sampling from the joint smoothing posterior p(x1,...xT | y1,...,yT) of the state space model with nonlinear measurement equation:

yₜ = C(xₜ) + εₜ, εₜ ~ N(0,Σₑ) Measurement equation

xₜ = Axₜ₋₁+ Buₜ + ηₜ, ηₜ ~ N(0,Σₙ) State equation

where

C(xₜ) is a non-linear function that we can ForwardDiff.jl to get the Jacobian

xₜ is the n-dim state

uₜ is the m-dim control

yₜ is the k-dim observed data.

The observed data observations are the rows of the T×k matrix Y The control signals are the rows of the T×m matrix U μ₀ and Σ₀ are the mean and covariance of the initial state vector x₀

Note: If nSim == 1, the returned Xdraws is matrix, otherwise it is a 3D array of size T×n×nIter.

SMCsamplers.FFBS_unscented — Function

FFBS_unscented(U, Y, A, B, C, Cargs, Σₑ, Σₙ, μ₀, Σ₀)

Forward filtering and backward sampling from the joint smoothing posterior p(x1,...xT | y1,...,yT) of the state space model with nonlinear measurement equation:

yₜ = C(xₜ) + εₜ, εₜ ~ N(0,Σₑ) Measurement equation

xₜ = Axₜ₋₁+ Buₜ + ηₜ, ηₜ ~ N(0,Σₙ) State equation

where

C(xₜ) is a non-linear function

xₜ is the n-dim state

uₜ is the m-dim control

yₜ is the k-dim observed data.

The observed data observations are the rows of the T×k matrix Y The control signals are the rows of the T×m matrix U μ₀ and Σ₀ are the mean and covariance of the initial state vector x₀

Note: If nSim == 1, the returned Xdraws is matrix, otherwise it is a 3D array of size T×n×nIter.

SMCsamplers.FFBS_laplace — Function

FFBS_laplace(U, Y, A, B, Σₙ, μ₀, Σ₀, observation, θ, nSim = 1; filter_output = false)

Forward filtering and backward sampling from the joint smoothing posterior p(x1,...xT | y1,...,yT) of the general state space model:

yₜ ~ p(yₜ | xₜ) Measurement model

xₜ ~ p(xₜ | xₜ₋₁) State transition model

The observed data observations are the rows of the T×k matrix Y The control signals are the rows of the T×m matrix U μ₀ and Σ₀ are the mean and covariance of the initial state vector x₀

Note: If nSim == 1, the returned Xdraws is matrix, otherwise it is a 3D array of size T×n×nIter.