SMCsamplers.kalmanfilter_update — Function

kalmanfilter_update(μ, Ω, u, y, A, B, C, Σₑ, Σₙ)

A single Kalman filter update at time t of the state space model:

yₜ = Cxₜ + εₜ, εₜ ~ N(0,Σₑ) Measurement equation xₜ = Axₜ₋₁+ Buₜ + ηₜ, ηₜ ~ N(0,Σₙ) State equation

where xₜ is the n-dim state uₜ is the m-dim control yₜ is the k-dim observed data.

Reference: Thrun, Burgard and Fox (2006). Probabilistic Robotics, Algorithm Kalman_filter in Table 3.

SMCsamplers.kalmanfilter_update_extended — Function

kalmanfilter_update_extended(μ, Ω, u, y, A, B, C, ∂C, Cargs, Σₑ, Σₙ)

A single extended Kalman filter update at time t of the state space model:

yₜ = C(xₜ) + εₜ, εₜ ~ N(0,Σₑ) Measurement equation xₜ = Axₜ₋₁+ Buₜ + ηₜ, ηₜ ~ N(0,Σₙ) State equation

where C(xₜ) is a non-linear measurement function

xₜ is the n-dim state uₜ is the m-dim control yₜ is the k-dim observed data.

Reference: Thrun, Burgard and Fox (2006). Probabilistic Robotics, Algorithm Kalman_filter in Table 3.

SMCsamplers.kalmanfilter_update_unscented — Function

kalmanfilter_update_unscented(μ, Ω, u, y, A, B, C, Cargs, Σₑ, Σₙ)

A single unscented Kalman filter update at time t of the state space model:

yₜ = C(xₜ) + εₜ, εₜ ~ N(0,Σₑ) Measurement equation xₜ = Axₜ₋₁+ Buₜ + ηₜ, ηₜ ~ N(0,Σₙ) State equation

where C(xₜ) is a non-linear measurement function

xₜ is the n-dim state uₜ is the m-dim control yₜ is the k-dim observed data.

Reference: Thrun, Burgard and Fox (2006). Probabilistic Robotics, Algorithm Kalman_filter in Table 3.4